feat(service): move graphql uploud logic to utils
